NCI-Agency/anet#1252: Add publish action to the ReportDao
We have added the publish action (including sending the needed
notification e-mail) to the ReportDao. This way it can also be used
later by the report publication worker.
